Add search functionality for desktop and mobile
There are currently two search forms: one in the header, and one in the search modal.

When the user submits a search form, the App component's action is called. This action redirects the user to a search results page, where the user's query is encoded in the URL.

The SearchResults component has a loader that retrieves the user's query from the URL. Then, the loader selects the exercises from civicPrototype.json that match that query.

Finally, these exercises are rendered in SearchResults.