Add exercise content ID to URL parameters
Previously, the exercise content ID was included in the URL as a hash fragment. The hash fragment was updated dynamically as the user pressed the sidebar buttons.

However, a user could not use this URL to go directly to a specific content item (e.g. by copy-pasting the link into their browser). Instead of the content item in the hash fragment, the user always saw the first item in the "Learn" tab.

Fix this issue by adding the content ID to the URL as a parameter instead of a hash fragment.

Example with exercise 2, content item learn1:
- Old URL: /exercises/2#learn1
- New URL: /exercises/2/learn1