Add categories to exercise dropdown; change width
Format category IDs in civicPrototype.json as an array, rather than a string. This will allow us to match categories with the exercises that belong to them.

Organize exercises in the dropdown by category. This change will make it easier for users to find relevant exercises.

With this new category layout, the dropdown requires more space. Expand the dropdown to fill the screen width on large screens.