Add error page; remove hard-coded navigation
If the user enters an invalid path or exercise ID, display a customized error page.

Unlike the default error page, this page retains the app's header, so the user can easily navigate back to the homepage or search for other content.

The ID of the first content item may not always be "learn1." Remove the hard-coded <Navigate>, which automatically set the content ID to "learn1."

As a result of this change, content IDs are now mandatory in every exercise page route. Failing to specify a content ID will send the user to the error page.