Improve search functionality
Add suggested queries and exercises to the website's search bars.

The queries are taken from the exercise tags in civicPrototype.json.

When a user clicks on a suggested exercise, they are taken directly to the exercise page instead of the search results page.

These changes should make it easier for users to find their desired content.